+++ app/models/recurring_task.rb
@@ -11,13 +11,17 @@ class RecurringTask < ActiveRecord::Base
   INTERVAL_MONTH = 'm'
   INTERVAL_YEAR = 'y'
   
-  # must come before validations otherwise unitialized
+  # must come before validations otherwise uninitialized
   INTERVAL_UNITS_LOCALIZED = [l(:interval_day), l(:interval_week), l(:interval_month), l(:interval_year)]
 
-  validates :interval_localized_name, presence: true, inclusion: { in: RecurringTask::INTERVAL_UNITS_LOCALIZED, message: "#{l(:error_invalid_interval)} '%{value}' (Validation)" }
-  validates :interval_number, presence: true, numericality: {only_integer: true, greater_than: 0}
-  # validates :issue, presence: true # cannot validate presence of issue if want to use other features
-  # validates :fixed_schedule # requiring presence requires true
+  # pulled out validates_presence_of separately
+  # for older Rails compatibility
+  validates_presence_of :interval_localized_name
+  validates_presence_of :interval_number
+  
+  validates :interval_localized_name, inclusion: { in: RecurringTask::INTERVAL_UNITS_LOCALIZED, message: "#{l(:error_invalid_interval)} '%{value}' (Validation)" }
+  validates :interval_number, numericality: {only_integer: true, greater_than: 0}
+  # cannot validate presence of issue if want to use other features; requiring presence of fixed_schedule requires it to be true
 
   validates_associated :issue # just in case we build in functionality to add an issue at the same time, verify the issue is ok  
   
@@ -33,8 +37,7 @@ class RecurringTask < ActiveRecord::Base
     when INTERVAL_YEAR
       l(:interval_year)
     else
-      logger.error "#{l(:error_invalid_interval)} #{interval_unit} (interval_localized_name)"
-      ""
+      raise "#{l(:error_invalid_interval)} #{interval_unit} (interval_localized_name)"
     end  
   end
   
@@ -56,8 +59,7 @@ class RecurringTask < ActiveRecord::Base
       when l(:interval_year)
         INTERVAL_YEAR
       else
-        logger.error "#{l(:error_invalid_interval)} #{value} (interval_localized_name=)"
-        ""
+        raise "#{l(:error_invalid_interval)} #{value} (interval_localized_name=)"
       end
   end
   
@@ -73,8 +75,7 @@ class RecurringTask < ActiveRecord::Base
     when INTERVAL_YEAR
       interval_number.years
     else
-      logger.error "#{l(:error_invalid_interval)} #{interval_unit} (recurrence_pattern)"
-      1.years
+      raise "#{l(:error_invalid_interval)} #{interval_unit} (recurrence_pattern)"
     end
   end